LayoutTests/printing/page-break-inside-expected.txt

Test for page-break-inside:avoid

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pageNumber('page1') is 1
PASS pageNumber('page2') is 2
PASS pageNumber('page3-1') is 3
PASS pageNumber('page3-2') is 3
PASS pageNumber('page4') is 4
PASS pageNumber('page5') is 5
PASS pageNumber('page6') is 6
PASS pageNumber('page7') is 7
PASS pageNumber('page9') is 9
PASS pageNumber('page10-1') is 10
PASS pageNumber('page10-2') is 10
PASS pageNumber('page11') is 11
PASS successfullyParsed is true

TEST COMPLETE
